Slimmed down centos container by removing python-related packages
The problem?

The centos container with the additional packages that we install
takes up about 410mb.

The solution!

Removed python-specific packages from this container to keep this
baseline container slim.  That way we can have two separate python
branches from this container, one for python27u & python36u and
one for pyenv